requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & APIs
httpx>=0.25.0
aiofiles>=23.2.0

# AI Services
google-generativeai>=0.3.0
elevenlabs>=1.0.0

# Content Aggregation
praw>=7.7.0
feedparser>=6.0.0

# Database & Storage
supabase>=2.0.0
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
pydub>=0.25.0
cryptography>=41.0.0

# ===== NEW: Agentic Content Intelligence =====

# Google AI (Primary)
google-genai>=1.0.0

# Semantic Processing (Local - FREE)
sentence-transformers>=3.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
hdbscan>=0.8.33

# Content Aggregation
newsdataapi>=0.1.0
youtube-transcript-api>=0.6.0

# Complementary Search (for counter-arguments)
exa-py>=1.0.0

# Email
resend>=2.0.0
email-validator>=2.1.0

# Audio Processing
google-cloud-texttospeech>=2.14.0
pydub>=0.25.0

# Database
psycopg2-binary>=2.9.0

# Dev
pytest>=7.4.0
pytest-asyncio>=0.21.0
